# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Inquiry {
  id: number;
  businessName: string;
  contactName: string;
  email: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  website: string | null;
  projectType: string;
  projectReason: string;
  referralSource: string;
  message: string;
  createdAt: string;
}

export interface InquiryInput {
  /** @minLength 2 */
  businessName: string;
  /** @minLength 2 */
  contactName: string;
  email: string;
  phone?: string;
  website?: string;
  /** @minLength 2 */
  projectType: string;
  /** @minLength 2 */
  projectReason: string;
  /** @minLength 2 */
  referralSource: string;
  /** @minLength 10 */
  message: string;
}

export interface Error {
  error: string;
}