# lib/api-zod/src/generated/types/inquiry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Inquiry {
  id: number;
  businessName: string;
  contactName: string;
  email: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  website: string | null;
  projectType: string;
  projectReason: string;
  referralSource: string;
  message: string;
  createdAt: Date;
}