# lib/api-zod/src/generated/types/inquiryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InquiryInput {
  /** @minLength 2 */
  businessName: string;
  /** @minLength 2 */
  contactName: string;
  email: string;
  phone?: string;
  website?: string;
  /** @minLength 2 */
  projectType: string;
  /** @minLength 2 */
  projectReason: string;
  /** @minLength 2 */
  referralSource: string;
  /** @minLength 10 */
  message: string;
}